requirements.txt
streamlit
pandas
numpy
scikit-learn
plotly
google-genai

